+++ cmd/gobench-cable/main.go
@@ -57,7 +57,7 @@ func main() {
 
 	controller := gobench.NewController(metrics)
 
-	appNode := node.NewNode(controller, metrics)
+	appNode := node.NewNode(controller, metrics, &config.App)
 	appNode.Start()
 
 	var disconnector node.Disconnector